refactor: add GetAttachmentsDownloadLinkProps type and reorganize type imports
