+++ src/types/index.d.ts
@@ -14,4 +14,3 @@ export * from './filters';
 export * from './messages';
 export * from './participant';
 export * from './soap';
-export * from './utils';
