+++ src/views/search/list/conversation/search-conversation-list-item-core.tsx
@@ -19,9 +19,10 @@ import { filter, forEach, includes, reduce, uniqBy } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
 import { ConversationSubjectRow } from '../../../app/folder-panel/parts/conversation-subject-row';
-import { NormalizedConversation, TextReadValuesProps } from 'types/index.d';
-import { ItemAvatar } from 'views/app/folder-panel/parts/item-avatar';
 import { ParticipantsString } from '../../../app/folder-panel/parts/participants-string';
+import { NormalizedConversation } from 'types/conversations';
+import { TextReadValuesProps } from 'types/utils';
+import { ItemAvatar } from 'views/app/folder-panel/parts/item-avatar';
 import { RowInfo } from 'views/app/folder-panel/parts/row-info';
 
 type SearchConversationListItemCoreProps = {
